requirements.txt
streamlit
langdetect
Pillow
pytesseract
transformers
torch
speechrecognition
pydub
requests
google-generativeai
absl-py
numpy
scikit-learn
spacy